backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
idna==3.10
psycopg2-binary==2.9.9
PyJWT==2.10.1
requests==2.32.3
setuptools==69.1.1
sqlparse==0.5.1
urllib3==2.2.3
wheel==0.43.0
